구매관리 기능 추가

+++ src/main/java/com/megazone/ERPSystem_phase2_Backend/logistics/model/purchase_management/PurchaseRequestDetail.java
@@ -4,6 +4,8 @@ import com.megazone.ERPSystem_phase2_Backend.logistics.model.product_registratio
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 /**
  * 발주요청 상세 테이블
  * 발주요청서에 포함된 품목과 관련된 정보 관리
@@ -36,19 +38,19 @@ public class PurchaseRequestDetail {
 
     // 단가(입고단가)
     @Column(nullable = false)
-    private Double price;
+    private BigDecimal price;
 
     // 공급가액 (수량 * 단가)
     @Column(nullable = false)
-    private Double supplyPrice;
+    private BigDecimal supplyPrice;
 
     // 원화금액 (통화가 외자일때만 사용)
     @Column(nullable = true)
-    private Double localAmount;
+    private BigDecimal localAmount;
 
     // 부가세
     @Column
-    private Double vat;
+    private BigDecimal vat;
 
     // 비고
     @Column
